# services/knowledge/api/openapi.yaml
openapi: 3.0.3
info:
  title: Knowledge Service Internal API
  version: 0.3.0
  description: >
    Service-local API contract for the Go-based Knowledge service. Public
    frontend contracts remain owned by the gateway OpenAPI. This service
    exposes operational endpoints and service-local /internal/v1 resources for
    knowledge metadata, document processing state, ingestion jobs, retrieval,
    runtime config, and statistics.
servers:
  - url: http://localhost:8000
...
paths:
...
  /internal/v1/knowledge-bases:
    get:
      tags: [knowledge-bases]
      summary: List knowledge bases
      operationId: listInternalKnowledgeBases
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
        - name: keyword
          in: query
          schema:
            type: string
        - name: docType
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Knowledge bases visible to the gateway user context.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeBaseListResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
    post:
      tags: [knowledge-bases]
      summary: Create knowledge base
      operationId: createInternalKnowledgeBase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateKnowledgeBaseRequest'
      responses:
        '201':
          description: Knowledge base created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeBaseResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /internal/v1/knowledge-bases/{knowledgeBaseId}:
    parameters:
      - name: knowledgeBaseId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [knowledge-bases]
      summary: Get knowledge base
      operationId: getInternalKnowledgeBase
      responses:
        '200':
          description: Knowledge base.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeBaseResponse'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    patch:
      tags: [knowledge-bases]
      summary: Update knowledge base
      operationId: updateInternalKnowledgeBase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateKnowledgeBaseRequest'
      responses:
        '200':
          description: Knowledge base updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeBaseResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    delete:
      tags: [knowledge-bases]
      summary: Delete knowledge base
      operationId: deleteInternalKnowledgeBase
      responses:
        '204':
          description: Knowledge base deleted.
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /internal/v1/knowledge-bases/{knowledgeBaseId}/documents:
    parameters:
      - name: knowledgeBaseId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [knowledge-bases]
      summary: List knowledge-base documents
      operationId: listInternalKnowledgeBaseDocuments
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/DocumentStatus'
      responses:
        '200':
          description: Documents in the knowledge base.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /internal/v1/knowledge-bases/{knowledgeBaseId}/ingestion-jobs:
    parameters:
      - name: knowledgeBaseId
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [knowledge-bases]
      summary: Create ingestion job from file handoff
      operationId: createInternalIngestionJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HandoffRequest'
      responses:
        '201':
          description: Ingestion job accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandoffResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /internal/v1/knowledge-bases/{knowledgeBaseId}/jobs:
    parameters:
      - name: knowledgeBaseId
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [knowledge-bases]
      summary: Create knowledge-base job
      operationId: createInternalKnowledgeBaseJob
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                jobType:
                  type: string
                  enum: [reprocess]
      responses:
        '202':
          description: Reprocessing job accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingJobResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /internal/v1/documents/{documentId}:
    parameters:
      - name: documentId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [knowledge-bases]
      summary: Get document processing details
      operationId: getInternalDocument
      responses:
        '200':
          description: Document processing details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /internal/v1/documents/{documentId}/chunks:
    parameters:
      - name: documentId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [knowledge-bases]
      summary: List document chunks
      operationId: listInternalDocumentChunks
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
      responses:
        '200':
          description: Document chunks.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentChunkListResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /internal/v1/jobs/{jobId}:
    parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [knowledge-bases]
      summary: Get processing job
      operationId: getInternalProcessingJob
      responses:
        '200':
          description: Processing job.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingJobResponse'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /internal/v1/jobs/{jobId}/processing-runs:
    parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [knowledge-bases]
      summary: Run an ingestion processing job
      operationId: createInternalProcessingRun
      responses:
        '201':
          description: Processing run completed for the current synchronous MVP implementation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingJobResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /internal/v1/knowledge-queries:
    post:
      tags: [knowledge-bases]
      summary: Create knowledge query
      operationId: createInternalKnowledgeQuery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeQueryRequest'
      responses:
        '201':
          description: Knowledge query result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeQueryResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '502':
          $ref: '#/components/responses/Error'
  /internal/v1/runtime-config:
    get:
      tags: [knowledge-bases]
      summary: Get sanitized runtime configuration
      operationId: getInternalRuntimeConfig
      responses:
        '200':
          description: Sanitized runtime configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeConfigResponse'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
    patch:
      tags: [knowledge-bases]
      summary: Update runtime configuration
      operationId: updateInternalRuntimeConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimeConfigUpdateRequest'
      responses:
        '200':
          description: Sanitized runtime configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeConfigResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
  /internal/v1/knowledge-stats:
    get:
      tags: [knowledge-bases]
      summary: Get knowledge statistics
      operationId: getInternalKnowledgeStats
      responses:
        '200':
          description: Knowledge statistics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeStatsResponse'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
components:
  responses:
    Error:
      description: Standard error envelope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    PageInfo:
      type: object
      required: [page, pageSize, total]
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
    ChunkStrategy:
      type: object
      additionalProperties: true
      properties:
        type:
          type: string
          example: SEMANTIC_TEXT
        chunkSize:
          type: integer
          minimum: 1
        overlap:
          type: integer
          minimum: 0
        separators:
          type: array
          items:
            type: string
...
    CreateKnowledgeBaseRequest:
      type: object
      required: [name]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        docType:
          type: string
        chunkStrategy:
          $ref: '#/components/schemas/ChunkStrategy'
        retrievalStrategy:
          $ref: '#/components/schemas/RetrievalStrategy'
    UpdateKnowledgeBaseRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        docType:
          type: string
        chunkStrategy:
          $ref: '#/components/schemas/ChunkStrategy'
        retrievalStrategy:
          $ref: '#/components/schemas/RetrievalStrategy'
...
    KnowledgeBaseResponse:
      type: object
      required: [data, requestId]
      properties:
        data:
          $ref: '#/components/schemas/KnowledgeBaseSummary'
        requestId:
          type: string
    KnowledgeBaseListResponse:
      type: object
      required: [data, page, requestId]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/KnowledgeBaseSummary'
        page:
          $ref: '#/components/schemas/PageInfo'
        requestId:
          type: string
    DocumentStatus:
      type: string
      enum:
        - uploaded
        - parsing
        - chunking
        - embedding
        - ready
        - failed
...
    DocumentResponse:
      type: object
      required: [data, requestId]
      properties:
        data:
          $ref: '#/components/schemas/DocumentSummary'
        requestId:
          type: string
    DocumentListResponse:
      type: object
      required: [data, page, requestId]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/DocumentSummary'
        page:
          $ref: '#/components/schemas/PageInfo'
        requestId:
          type: string
...
    DocumentChunkListResponse:
      type: object
      required: [data, page, requestId]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/DocumentChunk'
        page:
          $ref: '#/components/schemas/PageInfo'
        requestId:
          type: string
    HandoffRequest:
      type: object
      required: [fileId, name]
      properties:
        fileId:
          type: string
        name:
          type: string
        contentType:
          type: string
        sizeBytes:
          type: integer
          minimum: 0
        tags:
          type: array
          items:
            type: string
        createdBy:
          type: string
        idempotencyKey:
          type: string
...
    HandoffResponse:
      type: object
      required: [data, requestId]
      properties:
        data:
          $ref: '#/components/schemas/HandoffResult'
        requestId:
          type: string
...
    ProcessingJobResponse:
      type: object
      required: [data, requestId]
      properties:
        data:
          $ref: '#/components/schemas/ProcessingJob'
        requestId:
          type: string
    KnowledgeQueryRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
        knowledgeBaseIds:
          type: array
          items:
            type: string
        topK:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
        scoreThreshold:
          type: number
          minimum: 0
          default: 0
        tags:
          type: array
          items:
            type: string
        metadataFilter:
          type: object
          additionalProperties:
            type: string
        rerank:
          type: boolean
          default: false
        rerankTopN:
          type: integer
          nullable: true
          minimum: 1
...
    KnowledgeQueryResponse:
      type: object
      required: [data, requestId]
      properties:
        data:
          $ref: '#/components/schemas/KnowledgeQuerySummary'
        requestId:
          type: string
...
    RuntimeConfigUpdateRequest:
      type: object
      properties:
        parserBackend:
          type: string
        rerankProvider:
          type: string
        rerankModel:
          type: string
        retrievalTopK:
          type: integer
          minimum: 1
          maximum: 100
        scoreThreshold:
          type: number
          minimum: 0
        maxConcurrentJobs:
          type: integer
          minimum: 1
        processingTimeoutSec:
          type: integer
          minimum: 1
        secretRefs:
          type: object
          additionalProperties:
            type: string
    RuntimeConfigResponse:
      type: object
      required: [data, requestId]
      properties:
        data:
          $ref: '#/components/schemas/RuntimeConfig'
        requestId:
          type: string
...
    KnowledgeStatsResponse:
      type: object
      required: [data, requestId]
      properties:
        data:
          $ref: '#/components/schemas/KnowledgeStats'
        requestId:
          type: string